requirements.txt
bs4~=0.0.2
beautifulsoup4~=4.14.3
langchain-community~=0.4.1
langchain-text-splitters~=1.1.1
numpy~=2.4.3
soupsieve~=2.8.3
typing_extensions~=4.15.0
h11~=0.16.0
pip~=26.0.1
attrs~=26.1.0
protobuf~=6.33.6
filelock~=3.25.2
grpcio~=1.78.0
rich~=14.3.3
Pygments~=2.19.2
markdown-it-py~=4.0.0
requests~=2.33.0
PyYAML~=6.0.3
idna~=3.11
multidict~=6.7.1
propcache~=0.4.1
pydantic~=2.12.5
pydantic_core~=2.41.5
zipp~=3.23.0
importlib_metadata~=8.7.1
pyproject_hooks~=1.2.0
build~=1.4.2
click~=8.3.1
httpcore~=1.0.9
httpx~=0.28.1
zstandard~=0.25.0
mdurl~=0.1.2
charset-normalizer~=3.4.6
overrides~=7.7.0
sympy~=1.14.0
mpmath~=1.3.0
typer~=0.24.1
shellingham~=1.5.4
fsspec~=2026.2.0
aiohttp~=3.13.3
yarl~=1.23.0
MarkupSafe~=3.0.3
orjson~=3.11.7
PyPika~=0.51.1
uvloop~=0.22.1
xxhash~=3.6.0
aiosignal~=1.4.0
frozenlist~=1.8.0
aiohappyeyeballs~=2.6.1
certifi~=2026.2.25
urllib3~=2.6.3
uvicorn~=0.42.0
httptools~=0.7.1
websockets~=16.0
watchfiles~=1.1.1
chromadb~=1.5.5
importlib_resources~=6.5.2
pybase64~=1.4.3
tenacity~=9.1.4
packaging~=26.0
bcrypt~=5.0.0
onnxruntime~=1.24.4
jsonschema~=4.26.0
kubernetes~=35.0.0
mmh3~=5.2.1
anyio~=4.13.0
pydantic-settings~=2.13.1
tqdm~=4.67.3
six~=1.17.0
python-dateutil~=2.9.0.post0
oauthlib~=3.3.1
annotated-types~=0.7.0
langsmith~=0.7.22
uuid_utils~=0.14.1
requests-toolbelt~=1.0.0
langchain-core~=1.2.22
diskcache~=5.6.3
Jinja2~=3.1.6
websocket-client~=1.9.0
referencing~=0.37.0
jsonpointer~=3.1.1
jsonschema-specifications~=2025.9.1
durationpy~=0.10
requests-oauthlib~=2.0.0
SQLAlchemy~=2.0.48
tokenizers~=0.22.2
marshmallow~=3.26.2
flatbuffers~=25.12.19
jsonpatch~=1.33
huggingface_hub~=1.8.0
typing-inspect~=0.9.0
dataclasses-json~=0.6.7
python-dotenv~=1.2.2
mypy_extensions~=1.1.0
httpx-sse~=0.4.3
fastapi~=0.135.3
pymupdf